+++ src/context/context.test.ts
@@ -4,9 +4,9 @@ import { buildAuthHeaders } from "../auth";
 import { DH } from "../constants";
 import { findCfg, findInitialData } from ".";
 
-const id = process.env.MC_MSG_TEST_ID;
-const channelId = process.env.MC_MSG_TEST_CHANNEL_ID;
-const credentialsB64 = process.env.MC_MSG_TEST_CREDENTIALS_SECOND;
+const id = process.env.MC_TEST_VIDEO_ID;
+const channelId = process.env.MC_TEST_CHANNEL_ID;
+const credentialsB64 = process.env.MC_TEST_CREDENTIAL_2;
 const credentials = credentialsB64
   ? JSON.parse(Buffer.from(credentialsB64!, "base64").toString())
   : undefined;