+++ src/protobuf/assembler.test.ts
@@ -1,11 +1,13 @@
-import { gtsm, lrc } from "./assembler";
+import { transcriptFormatToken, liveReloadContinuation } from "./assembler";
 
 it("can generate lrc", () => {
-  expect(lrc({ videoId: "foo", channelId: "bar" }, { top: true })).toBe(
+  expect(
+    liveReloadContinuation({ videoId: "foo", channelId: "bar" }, { top: true })
+  ).toBe(
     "0ofMyAMxGihDZ3dxQ2dvRFltRnlFZ05tYjI4YUMrcW8zYmtCQlFvRFptOXZJQUU9MAGCAQIIBA%3D%3D"
   );
 });
 
 it("can generate gtsm", () => {
-  expect(gtsm("id", true)).toBe("CgNhc3ISAmVuGgA%3D");
+  expect(transcriptFormatToken("id", true)).toBe("CgNhc3ISAmVuGgA%3D");
 });